requirements.txt
requests
beautifulsoup4
pytz
Pillow